RadixSeekLowerBound initial buggy impl
This is against a known broken commit because I wanted to see if I could
find the bugs that we later fixed
(f63f49c0b598a5ead21c5015fb4d08fe7e3c21ea).

The result is: yes, we can. Given an tree built up of "111" and "112"
and a lower bound search of "11", the TLA spec crashes, and so does the
Go implementation (at the aforementioned commit), demonstrating the bug.
Woohoo!

I'm going to update the implementation to match HEAD next and verify
that we don't get any crashes.