DeletePrefix: add commented-out property that the tree is optimal
The DeletePrefix algorithm fails this check, but maybe that's okay. If
we wanted to ensure that DeletePrefix resulted in trees that remain
optimal, then the algorithm is wrong.

An "optimal" radix tree has the fewest nodes to get to a value for all
values in the tree. We can see in the debug output below that the tree
is not optimal after deleting a prefix.

I'm going to keep this spec in here commented out.

<< "tree check",
   "actual",
   [ Edges |->
         [ A |->
               [ Edges |->
                     [ B |->
                           [ Edges |-> << >>,
                             Prefix |-> <<"B">>,
                             Value |-> <<"A", "B">> ] ],
                 Prefix |-> <<"A">>,
                 Value |-> <<>> ] ],
     Prefix |-> <<>>,
     Value |-> <<>> ],
   "expected",
   [ Edges |->
         [ A |->
               [ Edges |-> << >>,
                 Prefix |-> <<"A", "B">>,
                 Value |-> <<"A", "B">> ] ],
     Prefix |-> <<>>,
     Value |-> <<>> ] >>